Refactor BinaryBuild architecture and add option to build from file
